// src/models/Disciplina.ts
export enum Disciplina {
  AprofundamentoEmLinguaPortuguesa = "APROFUNDAMENTO EM LINGUA PORTUGUESA",
  Biologia = "BIOLOGIA",
  EducacaoFisica = "EDUCAÇÃO FÍSICA",
  Filosofia = "FILOSOFIA",
  Fisica = "FÍSICA",
  Geografia = "GEOGRAFIA",
  GestaoDeStartupsI = "GESTÃO DE STARTUPS I",
  Historia = "HISTÓRIA",
  LinguaEspanhol = "LINGUA ESTRANGEIRA - ESPANHOL",
  LinguaIngles = "LINGUA ESTRANGEIRA - INGLES",
  LinguaPortuguesa = "LÍNGUA PORTUGUESA",
  Matematica = "MATEMÁTICA",
  NocoesDeRobotica = "NOÇÕES DE ROBÓTICA",
  ProgramacaoOrientadaAObjetos = "PROGRAMAÇÃO ORIENTADO A OBJETOS",
  ProgramacaoWeb = "PROGRAMAÇÃO WEB",
  Quimica = "QUÍMICA",
  SistemasOperacionais = "SISTEMAS OPERACIONAIS",
  Sociologia = "SOCIOLOGIA"
}