// src/models/Turma.ts
export enum Turma {
  Informatica1 = '1° Informática',
  Informatica2 = '2° Informática',
  Informatica3 = '3° Informática',

  Agropecuaria1 = '1° Agropecuária',
  Agropecuaria2 = '2° Agropecuária',
  Agropecuaria3 = '3° Agropecuária',

  Administracao1 = '1° Administração',
  Administracao2 = '2° Administração',
  Administracao3 = '3° Administração',

  Financas1 = '1° Finanças',
  Financas2 = '2° Finanças',
  Financas3 = '3° Finanças',
  
  EnergiasRenovaveis1 = '1° Energia Renovável',
  EnergiasRenovaveis2 = '2° Energia Renovável',
  EnergiasRenovaveis3 = '3° Energia Renovável',
}